+++ CHANGELOG.md
@@ -17,4 +17,6 @@
 
 - Add basic jwxt api
 
+- Add cache && modify api year and term
+